filters/isg/requirements.txt
biopython==1.70
numpy==1.14.0
